fix!: Kotlin server; Use Jackson3 serde by default.
