+++ .agents/skills/tachyon-mcp/resources/kotlin/ToolHandlerExample.kt
@@ -4,14 +4,13 @@ package dev.tachyonmcp.skill
 
 import dev.tachyonmcp.api.json.JsonSchema
 import dev.tachyonmcp.api.runtime.InteractionContext
-import dev.tachyonmcp.core.server.TachyonServer
 import dev.tachyonmcp.api.server.features.tools.ToolRequest
 import dev.tachyonmcp.api.server.features.tools.AbstractToolHandler
 import dev.tachyonmcp.api.server.features.tools.ToolResult
+import dev.tachyonmcp.kotlin.server.TachyonServer
 import dev.tachyonmcp.kotlin.server.buildServer
 import dev.tachyonmcp.kotlin.server.config.TachyonServerBuilder
 import dev.tachyonmcp.kotlin.server.domain.decode
-import dev.tachyonmcp.kotlin.server.features.tools.registerTool
 import dev.tachyonmcp.kotlin.server.features.tools.toolDescriptor
 import kotlinx.coroutines.delay
 import kotlinx.serialization.Serializable
