+++ examples/echo-kotlin/src/main/kotlin/com/example/echo/EchoServer.kt
@@ -5,9 +5,8 @@ package com.example.echo
 import dev.tachyonmcp.api.json.JsonSchema
 import dev.tachyonmcp.api.server.config.Mode
 import dev.tachyonmcp.api.server.features.tools.ToolResult
-import dev.tachyonmcp.core.server.TachyonServer
+import dev.tachyonmcp.kotlin.server.TachyonServer
 import dev.tachyonmcp.kotlin.server.buildServer
-import dev.tachyonmcp.kotlin.server.features.tools.registerTool
 
 fun assembleServer(port: Int = 0): TachyonServer {
     val boundPort = port
