+++ tachyon-kotlin/src/main/kotlin/dev/tachyonmcp/kotlin/server/config/JsonScope.kt
@@ -22,11 +22,8 @@ public class JsonScope
     internal constructor() {
         /**
          * Payload serializer/deserializer for structured values and arguments.
-         * `null` keeps the module-wide default
-         * ([dev.tachyonmcp.kotlin.server.json.KxSerializationSerde] in the Kotlin DSL).
+         * `null` keeps Tachyon's Jackson default.
          * Sets both [serializer] and [deserializer] when assigned.
-         * Assign a non-null value to override; the Kx default is set once at builder level,
-         * not as a side effect of opening this block.
          */
         public var serde: PayloadSerde? = null
