+++ tachyon-kotlin/src/test/kotlin/dev/tachyonmcp/kotlin/server/ToolFnFactoryTest.kt
@@ -10,7 +10,6 @@ import dev.tachyonmcp.core.server.TachyonServer
 import dev.tachyonmcp.core.server.internal.ServerEngine
 import dev.tachyonmcp.core.server.session.DefaultDispatchContext
 import dev.tachyonmcp.kotlin.server.features.CoroutineRuntime
-import dev.tachyonmcp.kotlin.server.features.tools.registerTool
 import dev.tachyonmcp.kotlin.server.features.tools.toolFn
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
